EZP-28038: Refined `ezurl` fieldtype
