--- lib/Form/Type/FieldValue/UrlType.php
@@ -1,57 +1,0 @@
-<?php
-/**
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- */
-namespace EzSystems\RepositoryForms\Form\Type\FieldValue;
-
-use Symfony\Component\Form\AbstractType;
-use Symfony\Component\Form\Extension\Core\Type\TextType;
-use Symfony\Component\Form\Extension\Core\Type\UrlType as BaseUrlType;
-use Symfony\Component\Form\FormBuilderInterface;
-use Symfony\Component\OptionsResolver\OptionsResolver;
-
-/**
- * Combined Type for ezurl.
- */
-class UrlType extends AbstractType
-{
-    public function getName()
-    {
-        return $this->getBlockPrefix();
-    }
-
-    public function getBlockPrefix()
-    {
-        return 'ezrepoforms_fieldtype_url';
-    }
-
-    public function buildForm(FormBuilderInterface $builder, array $options)
-    {
-        $builder
-            ->add(
-                'link',
-                BaseUrlType::class,
-                [
-                    'label' => 'content.field_type.ezurl.link',
-                    'required' => false,
-                ]
-            )
-            ->add(
-                'text',
-                TextType::class,
-                [
-                    'label' => 'content.field_type.ezurl.text',
-                    'required' => false,
-                ]
-            );
-    }
-
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver
-            ->setDefaults([
-                'translation_domain' => 'ezrepoforms_content',
-            ]);
-    }
-}
